+++ src/legacy-commands/uwuify.ts
@@ -19,7 +19,7 @@
 import type { Message } from 'discord.js';
 import Uwuifier from 'uwuifier';
 
-import type CommandConfig from '#lib/interfaces/commands/LexiCommandConfig';
+import type CommandConfig from '#lib/interfaces/commands/LegacyLexiCommandConfig';
 import LegacyLexiCommand from '#lib/structures/LegacyLexiCommand';
 
 const uwuify: Uwuifier = new Uwuifier();
